🎨 Consolidated posts test config into Vite
ref https://linear.app/ghost/issue/BER-3505/migrate-comment-moderation-filters-to-new-filter-primitives
The comments filter follow-up no longer needs a separate Vitest config, so the app test setup now lives in the main Vite config and the new tests were rewritten to rely on the existing @src pathing.
